Branch review: relocate middleware helpers, fix stale docstring paths
Reviewed every commit on this branch from main..HEAD (68 commits).
Almost all of them were clean. The fixes in this commit are:

1. Resurrect a small utils.py to hold functions that operate on middleware
   objects but don't take a Butler, moving them out of butlerQueries where
   the name was misdescriptive:

   - summaryStatsToDict (pure dict conversion of ExposureSummaryStats)
   - getAirmass (extracts airmass from an Exposure)
   - getExpIdOrVisitId (extracts an int from a DimensionRecord/DataCoordinate)

   butlerQueries is now strictly about helpers that call butler.registry,
   butler.collections or butler.dimensions, matching its module docstring.

2. Drop getNumExpectedItems from butlerQueries — dead code since bf9e9b68
   ("Use dispatched payloads for numExpected calculations"). No callers
   anywhere in python/, scripts/, or tests/.

3. Fix 24 stale `lsst.rubintv.production.utils.LocationConfig` docstring
   paths across 12 files (allSky, aos, catchupService, cleanup,
   oneOffProcessing, pipelineRunning, plotting/mosaicPlotting,
   plotting/mosaicing, plotting/nightReportPlotBase, rubinTv, starTracker,
   timedServices). LocationConfig now lives in locationConfig.py.
   a5ea1306 (the split that moved it) missed these references and the
   follow-up fd276928/27dbf841 passes only caught one of them.

4. Fix getExpIdOrVisitId docstring — it documented a parameter named
   `expRecord` but the actual argument is `obj` (and accepts both
   DimensionRecord and DataCoordinate). Now correct, fixed as part of the
   move to utils.py.

5. Fix raiseIf docstring in predicates.py — Raises block referenced
   `self.doRaise`, but raiseIf is a free function, not a method; and the
   exception name "AnyException" isn't a real type. Updated to describe
   the actual re-raise behaviour.

6. Fix getDoRaise Returns name — was `do_raise` (snake_case), now
   `doRaise` per project convention.

7. Add Returns blocks to runningCI / runningScons / runningPyTest in
   predicates.py. All three return bool but had no Returns section.

8. Fix timeFunction docstring example — referenced the old name
   `@timeFunc(logger)` instead of `@timeFunction(logger)`.

Concerns flagged but NOT fixed in this commit (each worth a separate
decision):

- c900bbf9 added an unconditional `raise ValueError` for any instrument
  other than LATISS/LSSTCam at the top of HeadProcessController
  .getPipelineConfig. scripts/LSSTComCam/runHeadNode.py and
  scripts/LSSTComCamSim/runHeadNode.py still exist and instantiate
  HeadProcessController; if either is ever deployed they will crash on
  the first exposure. Either the scripts should be deleted as dead code
  or the guard should be narrowed (moved inside the `case "cwfs"` branch,
  where the real instrument-specific logic lives). Needs a deployment-
  aware decision, not a mechanical fix.

- raiseIf (predicates.py) calls `sentry_sdk.capture_exception(error)`
  unconditionally, so every swallowed-with-doRaise=False exception still
  fires a Sentry event. There are ~33 callsites including routine
  expected-failure paths (catchupService, starTracker, timedServices,
  oneOffProcessing). May cause heavy Sentry noise. If desired, add an
  explicit `captureSentry` flag or only capture when `doRaise=True`.

- Every pod script calls `setupSentry()` before `setupLogging()`. Once
  980b4e7a added `logger.warning(...)` inside setupSentry, that warning
  now fires before the LSST logger is configured and may be silently
  dropped. Either swap the order in ~30 scripts or defer the warning
  until after logging is configured.

- Several extracted helpers have documentation gaps that are worth a
  dedicated pass but aren't bugs: makePlotFile / makePlotFileFromRecord
  have no docstrings; expRecordToUploadFilename doesn't document
  `extension` / `zeroPad`; parsers.getDimensionUniverse has no
  docstring, writeDimensionUniverseFile has a one-line placeholder;
  shardIo.getShardPath documents only one of three parameters;
  timing.py has no module docstring and DurationResult has no class
  docstring.

- formatters.getPodWorkerNumber is in formatters.py but isn't a
  formatter — it's env/argv plumbing. Fits startupChecks.py better.

- Several test files (test_metadataService.py, test_processingControl.py,
  test_timing.py) still have `"""Test cases for utils."""` as the
  module docstring — a pre-existing copy-paste that survived the split.

- Pre-existing typo: test class `CamearaControlConfigTestCase` in
  tests/test_processingControl.py (should be Camera…).

- 5b37a586 (aosRecipes extraction) claims "behaviour-preserving" but
  actually silently fixes two latent bugs — NaN was previously sent to
  ConsDB in a backfill path, and a `None:.2f` format crash path. Fixes,
  not regressions, but worth noting in case anyone is bisecting.

- 54ebac4f (ISR data-driven dispatch) similarly changes log message text
  despite claiming "no runtime behaviour change". Observable but harmless.

Commits reviewed (all 68 on main..HEAD):

01089c32 07150659 d3f5c901 676643b9 263b72d0 770385f1 119d5b4e 5086cdcd
cca6f29a 0884b80c 2b868d2e 11968056 b1d1c8c6 d56e915c 980b4e7a cde19d77
3ca8f3fb b40f23dd 0dd70cf1 a16a4e9d 72ac938f f4f38f38 159a93af 55acd6e0
3e984957 4984e7b1 c900bbf9 d39fd9a3 fe8c33b6 2307d039 4ee62068 8a35083c
a52118fa 0f612781 76293330 01abd18d 4a8de65c b4c267b4 6d08886b 813e6240
d76f2f9c 3617a6a0 50fb463f 610a2e05 1cb7b3ad a5ea1306 7d2f42e8 5f508b44
18105ac3 27dbf841 1885925b fd276928 430af920 7f4b2c3e e236252d 44f528f7
5b37a586 e5f4af70 de8b2faf 54ebac4f b5f7dbf0 3311e344 82b69345 3e4a8f30
89e944a9 40fe6e6c ba955224 5b7a5253

Verified: all 204 unit tests pass; flake8 clean.
