+++ python/lsst/rubintv/production/processingControl.py
@@ -52,7 +52,6 @@ from lsst.pipe.base import Instrument, Pipeline, PipelineGraph, TaskFactory
 from lsst.utils import getPackageDir
 from lsst.utils.packages import Packages
 
-from .butlerQueries import getExpIdOrVisitId
 from .locationConfig import LocationConfig
 from .payloads import Payload, pipelineGraphToBytes
 from .podDefinition import PodDetails, PodFlavor
@@ -60,6 +59,7 @@ from .predicates import isCalibration, isWepImage, runningCI
 from .redisUtils import ExposureProcessingInfo, RedisHelper
 from .shardIo import getShardPath, writeExpRecordMetadataShard, writeMetadataShard
 from .timing import BoxCarTimer
+from .utils import getExpIdOrVisitId
 
 if TYPE_CHECKING:
     from numpy.typing import NDArray
